+++ packages/stdlib/_src/collections/Array/from.ts
@@ -1,5 +1,5 @@
 /**
- * @tsplus getter Collection asArray
+ * @tsplus getter Collection toArray
  */
 export function from<A>(data: Collection<A>): Array<A> {
   return Array.from(data)