+++ packages/stdlib/_src/collections/Chunk/asArrayLike.ts
@@ -3,8 +3,8 @@ import { concreteChunkId } from "@tsplus/stdlib/collections/Chunk/definition"
 /**
  * Converts a chunk to an `ArrayLike` (either `Array` or `Buffer`).
  *
- * @tsplus getter Chunk asArrayLike
+ * @tsplus getter Chunk toArrayLike
  */
-export function asArrayLike<A>(self: Chunk<A>): ArrayLike<A> {
+export function toArrayLike<A>(self: Chunk<A>): ArrayLike<A> {
   return concreteChunkId(self)._arrayLike()
 }