+++ packages/stdlib/_src/collections/HashMap/asCollection.ts
@@ -1,6 +1,6 @@
 /**
- * @tsplus getter HashMap asCollection
+ * @tsplus getter HashMap toCollection
  */
-export function asCollection<K, V>(self: HashMap<K, V>): Collection<Tuple<[K, V]>> {
+export function toCollection<K, V>(self: HashMap<K, V>): Collection<Tuple<[K, V]>> {
   return self
 }