+++ packages/stdlib/_src/collections/ImmutableArray/from.ts
@@ -1,6 +1,6 @@
 /**
  * @tsplus static ImmutableArray/Ops from
- * @tsplus getter Collection asImmutableArray
+ * @tsplus getter Collection toImmutableArray
  */
 export function from<A>(iterable: Collection<A>): ImmutableArray<A> {
   return new ImmutableArray(Array.from(iterable))