+++ packages/stdlib/_src/collections/List/toSet.ts
@@ -1,6 +1,0 @@
-/**
- * @tsplus getter List toSet
- */
-export function toSet<A>(self: List<A>): Set<A> {
-  return new Set(self)
-}